+++ RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # RELEASE NOTES
 
+## 3.1.0
+
+* Implemented basic mixin support in NodeImpl
+* Implemented basic NamePathResolver and NamespaceRegistry support in SessionImpl
+
 ## 3.0.0
 
 Updated to support AEM 6.0, from AEM 5.6.1